'bug' fix to ii_driftcorrect: allows for missing selections
Previously, we had set up iI_driftcorrect to allow for several modes,
one (the one we use) required an identical # of selections to trials in
order to function. For one subj (AY), some trials are missing
selections. I adjusted the functionality so that, for each selection,
we look up the trial # and use that to select the entire trial for
drift correction. Potentially in the future we can update, or rename
this function ii_driftcorrectbytrial or something?