fixed a bug in ii_selectfixationsbytrial
The previous implementation of the buffer window for 'last' fixation selection in ii_selectfixationsbytrial (used for calibration) was doubling the duration, since I think it was assuming input was in samples, not ms. Now, we convert input (ms) to samples, and all works a bit better. Now using 'last' mode for demo.