updates to ii_smooth: dynamic Gaussian kernel size
Previously, we were using a fixed convolution kernel width of 101 samples (-50:50), regardless of the width of the Gaussian kernel embedded within it. When the wide kernel is convolved with time-series data containing NaNs, the NaNs are pushed forward/back by the width of the kernel (/2). I've updated the Gaussian mode to use a convolution kernel that's 3x the smoothing amount wide in each direction, which should minimize the NaN issue.

Potentially, in the future we may include an option to run smoothing on data interpolated over blinks/censored timepoints. For now, extended NaN epoch should be reasonably minimized.