added 0's to polyfit inputs to 'anchor' recalibration to origin
Since we drift correct before we recalibrate, it's sensical to enfore that 0,0 be a reference point for the best-fit polynomial. We add n_trials 0,0 pairs to the inputs to polyfit, where n_trials is the number of points used to recalibrate (# of trials that pass inclusion criterion). Potentially future versions may fit a polynomial which definitively passes through the origin (not necessarily enforced here).